+++ src/documents/apps.py
@@ -12,8 +12,8 @@ class DocumentsConfig(AppConfig):
         from documents.signals import document_updated
         from documents.signals.handlers import add_inbox_tags
         from documents.signals.handlers import add_to_index
-        from documents.signals.handlers import run_workflow_added
-        from documents.signals.handlers import run_workflow_updated
+        from documents.signals.handlers import run_workflows_added
+        from documents.signals.handlers import run_workflows_updated
         from documents.signals.handlers import set_correspondent
         from documents.signals.handlers import set_document_type
         from documents.signals.handlers import set_log_entry
@@ -27,7 +27,7 @@ class DocumentsConfig(AppConfig):
         document_consumption_finished.connect(set_storage_path)
         document_consumption_finished.connect(set_log_entry)
         document_consumption_finished.connect(add_to_index)
-        document_consumption_finished.connect(run_workflow_added)
-        document_updated.connect(run_workflow_updated)
+        document_consumption_finished.connect(run_workflows_added)
+        document_updated.connect(run_workflows_updated)
 
         AppConfig.ready(self)
